requirements.txt
appdirs==1.4.0
Django==1.10.5
django-analytical==2.2.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-jet==1.0.5
django-seo2==1.0.1
django-summernote==0.8.6
django-taggit==0.22.0
django-taggit-helpers==0.1.4
django-taggit-labels==0.4.1
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.10.3
oauth2client==4.0.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.1.10
rsa==3.4.2
six==1.10.0
sqlparse==0.2.2
uritemplate==3.0.0
